requirements.txt
Pillow
numpy
numba
